land-deals-backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Compress==1.15
mysql-connector-python==8.2.0
PyJWT==2.8.0
reportlab==4.0.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
bcrypt==4.1.2
